add content manager listener when create tool window content
