--- PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-eventhubs/src/main/java/com/microsoft/azure/toolkit/intellij/eventhubs/view/EventHubsToolWindowFactory.java
@@ -1,28 +1,0 @@
-/*
- * Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- */
-
-package com.microsoft.azure.toolkit.intellij.eventhubs.view;
-
-import com.intellij.icons.AllIcons;
-import com.intellij.openapi.project.Project;
-import com.intellij.openapi.wm.ToolWindow;
-import com.intellij.openapi.wm.ToolWindowFactory;
-import org.jetbrains.annotations.NotNull;
-
-public class EventHubsToolWindowFactory implements ToolWindowFactory {
-    @Override
-    public void init(@NotNull ToolWindow toolWindow) {
-        toolWindow.setToHideOnEmptyContent(true);
-    }
-
-    @Override
-    public boolean shouldBeAvailable(@NotNull Project project) {
-        return false;
-    }
-
-    @Override
-    public void createToolWindowContent(@NotNull Project project, @NotNull ToolWindow toolWindow) {
-    }
-}
