+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-eventhubs/src/main/java/com/microsoft/azure/toolkit/intellij/eventhubs/view/EventHubsToolWindowManager.java
@@ -7,6 +7,7 @@ package com.microsoft.azure.toolkit.intellij.eventhubs.view;
 
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.wm.ToolWindow;
+import com.intellij.openapi.wm.ToolWindowFactory;
 import com.intellij.openapi.wm.ToolWindowManager;
 import com.intellij.ui.content.*;
 import com.microsoft.azure.toolkit.lib.Azure;
@@ -25,16 +26,16 @@ import java.util.Optional;
 public class EventHubsToolWindowManager {
     private static final String EVENT_HUBS_TOOL_WINDOW = "Azure Event Hubs";
     private static final EventHubsToolWindowManager instance = new EventHubsToolWindowManager();
-    private final BidiMap<String, String> resourceIdToNameMap = new DualHashBidiMap<>();
+    private static final BidiMap<String, String> resourceIdToNameMap = new DualHashBidiMap<>();
 
     public static  EventHubsToolWindowManager getInstance() {
         return instance;
     }
 
     public void showEventHubsPanel(Project project, EventHubsInstance instance, boolean isListening) {
-        final ToolWindow toolWindow = getToolWindow(project);
+        final ToolWindow toolWindow =  ToolWindowManager.getInstance(project).getToolWindow(EVENT_HUBS_TOOL_WINDOW);
         final String contentName = getConsoleViewName(instance.getId(), instance.getName());
-        Content content = toolWindow.getContentManager().findContent(contentName);
+        Content content = Optional.ofNullable(toolWindow).map(w -> w.getContentManager().findContent(contentName)).orElse(null);
         if (content == null) {
             final EventHubsSendListenPanel panel = new EventHubsSendListenPanel(project, instance);
             content = ContentFactory.getInstance().createContent(panel, contentName, false);
@@ -50,9 +51,9 @@ public class EventHubsToolWindowManager {
     }
 
     public void stopListening(Project project, EventHubsInstance instance) {
-        final ToolWindow toolWindow = getToolWindow(project);
+        final ToolWindow toolWindow = ToolWindowManager.getInstance(project).getToolWindow(EVENT_HUBS_TOOL_WINDOW);
         final String contentName = getConsoleViewName(instance.getId(), instance.getName());
-        final Content content = toolWindow.getContentManager().findContent(contentName);
+        final Content content = Optional.ofNullable(toolWindow).map(w -> w.getContentManager().findContent(contentName)).orElse(null);
         if (Objects.isNull(content)) {
             return;
         }
@@ -75,35 +76,46 @@ public class EventHubsToolWindowManager {
         return result;
     }
 
-    private ToolWindow getToolWindow(Project project) {
-        final ToolWindow toolWindow = ToolWindowManager.getInstance(project).getToolWindow(EVENT_HUBS_TOOL_WINDOW);
-        Optional.ofNullable(toolWindow).ifPresent(w -> w.getContentManager().addContentManagerListener(new ContentManagerListener() {
-            @Override
-            public void contentRemoved(@NotNull ContentManagerEvent event) {
-                final String displayName = event.getContent().getDisplayName();
-                final String removeResourceId = resourceIdToNameMap.getKey(displayName);
-                Optional.ofNullable(removeResourceId).ifPresent(r -> {
-                    final EventHubsInstance instance = Azure.az(AzureEventHubsNamespace.class).getById(r);
-                    Optional.ofNullable(instance).ifPresent(EventHubsInstance::stopListening);
-                    resourceIdToNameMap.removeValue(displayName);
-                });
-            }
+    public static class EventHubsToolWindowFactory implements ToolWindowFactory {
+        @Override
+        public void init(@NotNull ToolWindow toolWindow) {
+            toolWindow.setToHideOnEmptyContent(true);
+        }
 
-            @Override
-            public void contentRemoveQuery(@NotNull ContentManagerEvent event) {
-                final String displayName = event.getContent().getDisplayName();
-                final boolean canClose = AzureMessager.getMessager().confirm(AzureString.format(
-                        "This will stop listening to \"{0}\", are you sure to do this?", displayName));
-                if (!canClose) {
-                    event.consume();
-                } else {
-                    final JComponent contentComponent = event.getContent().getComponent();
-                    if (contentComponent instanceof EventHubsSendListenPanel) {
-                        ((EventHubsSendListenPanel) contentComponent).dispose();
+        @Override
+        public boolean shouldBeAvailable(@NotNull Project project) {
+            return false;
+        }
+
+        @Override
+        public void createToolWindowContent(@NotNull Project project, @NotNull ToolWindow toolWindow) {
+            toolWindow.getContentManager().addContentManagerListener(new ContentManagerListener() {
+                @Override
+                public void contentRemoved(@NotNull ContentManagerEvent event) {
+                    final String displayName = event.getContent().getDisplayName();
+                    final String removeResourceId = resourceIdToNameMap.getKey(displayName);
+                    Optional.ofNullable(removeResourceId).ifPresent(r -> {
+                        final EventHubsInstance instance = Azure.az(AzureEventHubsNamespace.class).getById(r);
+                        Optional.ofNullable(instance).ifPresent(EventHubsInstance::stopListening);
+                        resourceIdToNameMap.removeValue(displayName);
+                        final JComponent contentComponent = event.getContent().getComponent();
+                        if (contentComponent instanceof EventHubsSendListenPanel) {
+                            ((EventHubsSendListenPanel) contentComponent).dispose();
+                        }
+                    });
+                }
+
+                @Override
+                public void contentRemoveQuery(@NotNull ContentManagerEvent event) {
+                    final String displayName = event.getContent().getDisplayName();
+                    final boolean canClose = AzureMessager.getMessager().confirm(AzureString.format(
+                            "This will stop listening to Event Hubs \"{0}\", are you sure to do this?", displayName));
+                    if (!canClose) {
+                        event.consume();
                     }
                 }
-            }
-        }));
-        return toolWindow;
+            });
+        }
     }
+
 }
