+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-eventhubs/src/main/resources/META-INF/azure-intellij-plugin-eventhubs.xml
@@ -7,7 +7,7 @@
     <extensions defaultExtensionNs="com.intellij">
     <toolWindow anchor="bottom"
                 icon="/icons/Common/event_hubs.svg"
-                factoryClass="com.microsoft.azure.toolkit.intellij.eventhubs.view.EventHubsToolWindowFactory"
+                factoryClass="com.microsoft.azure.toolkit.intellij.eventhubs.view.EventHubsToolWindowManager$EventHubsToolWindowFactory"
                 id="Azure Event Hubs" canCloseContents="true"/>
     </extensions>
 </idea-plugin>
